refactor: RunMain 파일 분리(Todo, User)
